Refactor ChatbotWidget into smaller sub-components and hooks
- Extracted state and logic into custom hooks (`useChatbotSettings`, `useSelectionAI`, `useVoiceRecognition`, `useChatStreaming`).
- Extracted atomic UI components (`ChatHeader`, `FloatingActionButton`, `ContextualAIButton`, `OfflineState`, `ToolSelector`, `ModelSelector`, `VoiceInputControl`).
- Extracted message rendering components (`ToolCard`, `StepHistory`, `MdContent`, `MessageList`).
- Combined input components into `ChatInput`.
- Cleaned up `ChatbotWidget.js` to act as a container orchestrating the hooks and components.
