+++ src/components/chatbot/MessageList.js
@@ -1,0 +1,106 @@
+import { User, Bot } from 'lucide-react';
+import ToolCard from './ToolCard';
+import StepHistory from './StepHistory';
+import MdContent from './MdContent';
+
+export default function MessageList({
+  messages,
+  isLoading,
+  messagesEndRef,
+  handleUIInteract,
+  handleLinkClick,
+}) {
+  return (
+    <div className="flex-1 overflow-y-auto p-3 sm:p-5 space-y-3 bg-gradient-to-b from-white/50 to-neutral-50/50 custom-chat-scrollbar">
+      {messages.map((message, index) => {
+        if (message.role === 'tool_action') {
+          if (message.done) return null;
+          return (
+            <div
+              key={message.id}
+              className="flex gap-3 justify-start animate-in slide-in-from-bottom-2 duration-300 w-full"
+            >
+              <div className="w-7 h-7 shrink-0 mt-1" />
+              <div className="flex-1 min-w-0 max-w-[95%]">
+                <ToolCard label={message.label} Icon={message.Icon} pending={true} />
+              </div>
+            </div>
+          );
+        }
+
+        if (message.hidden) return null;
+        const isAssistant = message.role === 'assistant';
+
+        let isConsecutive = false;
+        for (let i = index - 1; i >= 0; i--) {
+          const prevMsg = messages[i];
+          if (prevMsg.hidden) continue;
+          if (prevMsg.role === 'tool_action' && prevMsg.done) continue;
+          isConsecutive = prevMsg.role === message.role;
+          break;
+        }
+
+        return (
+          <div
+            key={message.id}
+            className={`flex gap-3 ${message.role === 'user' ? 'flex-row-reverse' : 'flex-row'} animate-in slide-in-from-bottom-2 duration-300 w-full`}
+          >
+            <div className="relative w-7 h-7 shrink-0 mt-1 flex items-center justify-center">
+              {!isConsecutive && (
+                <div
+                  className={`w-full h-full rounded-full flex items-center justify-center shadow-sm ${message.role === 'user' ? 'bg-black ring-2 ring-black/10' : 'bg-neutral-100 border border-neutral-200/60'}`}
+                >
+                  {message.role === 'user' ? (
+                    <User className="w-3.5 h-3.5 text-white" />
+                  ) : (
+                    <Bot className="w-3.5 h-3.5 text-neutral-600" />
+                  )}
+                </div>
+              )}
+            </div>
+
+            <div
+              className={`flex-1 min-w-0 ${message.role === 'user' ? 'max-w-[85%]' : 'max-w-[95%]'} flex flex-col ${message.role === 'user' ? 'items-end' : 'items-start'}`}
+            >
+              {isAssistant && message.steps?.length > 0 && (
+                <StepHistory
+                  steps={message.steps}
+                  onInteract={(block, action) => handleUIInteract(message.id, block, action)}
+                />
+              )}
+
+              {message.content && (
+                <div
+                  className={`px-3.5 sm:px-4 py-2.5 sm:py-3 rounded-2xl shadow-sm text-[13px] overflow-hidden ${message.role === 'user' ? 'bg-gradient-to-br from-black to-neutral-900 text-white shadow-black/20 rounded-tr-sm' : 'bg-white/90 backdrop-blur-sm text-neutral-900 shadow-neutral-200/50 border border-neutral-200/50 rounded-tl-sm'}`}
+                >
+                  {message.role === 'assistant' ? (
+                    <MdContent content={message.content} onLinkClick={handleLinkClick} />
+                  ) : (
+                    <MdContent content={message.content} isUser={true} />
+                  )}
+                  <p className={`text-[10px] mt-1.5 text-neutral-400`}>
+                    {message.timestamp.toLocaleTimeString([], {
+                      hour: '2-digit',
+                      minute: '2-digit',
+                    })}
+                  </p>
+                </div>
+              )}
+            </div>
+          </div>
+        );
+      })}
+
+      {isLoading && !messages.some((m) => m.role === 'tool_action' && !m.done) && (
+        <div className="flex gap-3 flex-row animate-in slide-in-from-bottom-2 duration-300 w-full mt-1">
+          <div className="relative w-7 h-7 shrink-0 flex items-center justify-center">
+            <div className="absolute inset-0 rounded-full border-[1.5px] border-neutral-200" />
+            <div className="absolute inset-0 rounded-full border-[1.5px] border-transparent border-t-neutral-800 border-r-neutral-800 animate-spin" />
+            <Bot className="w-3.5 h-3.5 text-neutral-600" />
+          </div>
+        </div>
+      )}
+      <div ref={messagesEndRef} />
+    </div>
+  );
+}
